Change to Apache 2.0 License
Large parts of this code base was copied directly from a PR by @emanlove in the Selenium2Library repository. S2L is written under the Apache 2.0 License and thus the governing license for this library should've also been set to Apache 2.0. The introduction of an MIT license was an unintentionally mistake.