Added implicit wait for angular to find elements.
Just as protractor applies wait for angular on the first find all elements and
with almost all AngularJS function calls, the AngularJSLibrary now adds an
implicit wait. Before each find we wait for angular. By placing the wait for
angular method within the find element the AngularJSLibrary implements similar
wait before each selenium call functionality due to the fact the Selenium2Library all
keywords must either find the element given a locator or take a WebElement.
The current implimentation does not yet wait for angular when WebElements are
given.

ToDo - Wait for angular when WebElements are used
ToDo - Add a flag to be able to turn on/off the implicit wait for angular
       functionality within AngularJSLibrary.