Added flag allowing for toggling implicit angular wait.
Added argument when importing AngularJSLibrary which allows one to turn off
or ignore the implicit wait for angular functionality. By default the flag is
set to False and thus the AngularJS library will wait for any outstanding
angular $http requests and $timeouts each time the Selenium2Library finds
an element by locator. The wait for angular functionality is not yet
implemented for WebElements. Also added was a test case testing the setting
of the ignore flag.