Removed confusing Angular/JavaScript syntax
The JavaScript code uses the dollar sign as syntax. Due to the usage
of the same symbol in Robot Framework the keyword documentation,
which contained some JavaScript syntax and dollar signs, there was
some confusion. The JavaScript, in most instances, was not adding to
the documentation and was thus removed.