extended errormessage root selector
The error that is trown when the default rootselector ng-app is used, is not realy clear about adjusting the way the library is imported:
"Keyword 'Element Text Should Be' failed after retrying for 30 seconds. The last error was: WebDriverException: Message: unknown error: Unable to find root selector using "[ng-app]". Please refer to the AngularJS library documentation for more information on how to resolve this error."

I made a smal adjustment to add: Unable to find root selector *that is given by importing the library* using