Clarrifying the status of the AngularJSLibrary
It has been noted that some feel this library is no longer "useful" as
there have been very little changes of late. Instead it is because this
is mature and stable library providing a very specific functionality,
provide it well, and has very little need for change. Thus I am updating
the metadata on this project to reflect this reality.

I think I may need to say as such within the project documentation but
that can come in another commit.